Live Log wall created but seems to be a bit slow, should we add multi-threading, next step! Create like has bugs, also change the feed for like and use case to plain text

+++ src/dilmaj/shared/MessageComposite.java
@@ -6,7 +6,7 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class MessageComposite implements IsSerializable {
 	String message;
-	Long sequence;
+	int sequence;
 	
 	public MessageComposite() {
 		
@@ -20,11 +20,11 @@ public class MessageComposite implements IsSerializable {
 		return message;
 	}
 
-	public void setSequence(long sequence) {
+	public void setSequence(int sequence) {
 		this.sequence=sequence;
 	}
 	
-	public long getSequence() {
+	public int getSequence() {
 		return sequence;
 	}
 }
